refactor code according to requirements
