+++ packages/react/src/components/form/form.tsx
@@ -1,7 +1,3 @@
-import "@fontsource/inter/400.css";
-import "@fontsource/inter/500.css";
-import "@fontsource/inter/600.css";
-
 import { clsx } from "clsx";
 import { useFlowState } from "./useFlowState";
 import { Initial } from "./initial";
