--- packages/react/src/components/index.ts
@@ -1,4 +1,0 @@
-import { Form } from "./form/form";
-import { ConfigurationProvider } from "../context/config-context";
-
-export { Form, ConfigurationProvider };
