+++ packages/react/src/main.ts
@@ -5,7 +5,7 @@ import {
 } from "./context/config-context";
 import { LoggedIn } from "./components/logged-in";
 import { LoggedOut } from "./components/logged-out";
-import { Form } from "./components";
+import { Form } from "./components/form";
 import { useSlashID } from "./hooks/use-slash-id";
 
 /**
